requirements/dev.txt
apscheduler==3.6.3
    # via python-telegram-bot
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   outcome
    #   pytest
    #   trio
backcall==0.2.0
    # via ipython
black==21.11b1
    # via -r requirements/dev.in
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via
    #   python-telegram-bot
    #   urllib3
cffi==1.15.0
    # via cryptography
click==8.0.3
    # via
    #   black
    #   pip-tools
cryptography==36.0.0
    # via
    #   pyopenssl
    #   urllib3
decorator==5.1.0
    # via
    #   ipdb
    #   ipython
flake8==4.0.1
    # via -r requirements/dev.in
h11==0.12.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements/dev.in
ipython==7.30.0
    # via ipdb
isort==5.10.1
    # via -r requirements/dev.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
mypy==0.910
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
outcome==1.1.0
    # via trio
packaging==21.3
    # via pytest
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.4.0
    # via -r requirements/dev.in
platformdirs==2.4.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.23
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.4.0
    # via flake8
pygments==2.10.0
    # via ipython
pyopenssl==21.0.0
    # via urllib3
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via -r requirements/dev.in
python-telegram-bot==13.8.1
    # via -r requirements/prod.in
pytz==2021.3
    # via
    #   apscheduler
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
regex==2021.11.10
    # via black
selenium==4.1.0
    # via -r requirements/dev.in
six==1.16.0
    # via
    #   apscheduler
    #   pyopenssl
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
toml==0.10.2
    # via
    #   ipdb
    #   mypy
    #   pytest
tomli==1.2.2
    # via
    #   black
    #   pep517
tornado==6.1
    # via python-telegram-bot
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
types-requests==2.26.1
    # via -r requirements/dev.in
typing-extensions==4.0.1
    # via
    #   black
    #   mypy
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via apscheduler
urllib3[secure]==1.26.7
    # via selenium
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
wsproto==1.0.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
